+++ src/multidim/intervalbox.jl
@@ -8,6 +8,7 @@ if VERSION >= v"0.6.0-dev"
     immutable IntervalBox{N,T} <: StaticVector{N, Interval{T}}
         data::NTuple{N,Interval{T}}
     end
+
 else
     doc"""An `IntervalBox` is an $N$-dimensional rectangular box, given
     by a Cartesian product of $N$ `Interval`s.