+++ test/interval_tests/rounding.jl
@@ -1,23 +1,20 @@
 using IntervalArithmetic
 using Base.Test
 
-# using Suppressor
 
 setformat(:full)
-
-# @suppress begin
 
 # @testset "Interval rounding" begin
 
 # NB: Due to "world age" problems, the following is not a @testset
 
-setrounding(Interval, :correct)
+setrounding(Interval, :slow)
 x = Interval(0.5)
 @testset "Correct rounding" begin
     @test sin(x) == Interval(0.47942553860420295, 0.479425538604203)
 end
 
-setrounding(Interval, :fast)
+setrounding(Interval, :accurate)
 @testset "Fast rounding" begin
     @test sin(x) == Interval(0.47942553860420295, 0.47942553860420306)
 end
@@ -27,8 +24,13 @@ setrounding(Interval, :none)
     @test sin(x) == Interval(0.479425538604203, 0.479425538604203)
 end
 
-setrounding(Interval, :correct)
+setrounding(Interval, :slow)
 @testset "Back to correct rounding" begin
+    @test sin(x) == Interval(0.47942553860420295, 0.479425538604203)
+end
+
+setrounding(Interval, :tight)
+@testset "Back to error-free rounding" begin
     @test sin(x) == Interval(0.47942553860420295, 0.479425538604203)
 end